code/requirements.txt
flask
flask-cors
gunicorn
numpy
pandas
bcrypt
google-cloud-firestore
google-cloud-bigquery
